+++ src/npc/ai/intro.rs
@@ -122,7 +122,7 @@ impl NPC {
 
         self.anim_counter += 1;
         if (self.anim_counter % 8) == 1 {
-            state.create_caret(self.x + state.effect_rng.range(-8..8) as isize * 0x200,
+            state.create_caret(self.x + state.effect_rng.range(-8..8) as i32 * 0x200,
                                self.y + 8 * 0x200,
                                CaretType::LittleParticles, Direction::Up);
         }
